requirements.txt
# Core dependencies voor RAG API
supabase==2.3.4
python-dotenv==1.0.1

# LangChain & Gemini
langchain==0.1.9
langchain-core==0.1.27
langchain-community==0.0.24
langchain-google-genai==0.0.9
google-generativeai==0.3.2

# Pydantic voor structured output
pydantic==2.6.1

# Streamlit frontend
streamlit==1.31.0
requests==2.31.0
plotly==5.18.0
pandas==2.0.3

# Vector search & embeddings
numpy==1.24.3
